backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.5.2
Flask-Cors==3.0.10
PyMySQL==1.1.0
python-dotenv==1.0.0